requirements.txt
fastapi
uvicorn
boto3
joblib
scikit-learn
python-dotenv
pandas
SQLAlchemy
imbalanced-learn